Couple bug fixes, stricmp crap, & version number
Fixed a case where the compiler would report the incorrect file and line
number for a OBJ keyword not in the first column.
Made another attempt at resolving the madness that is stricmp/strcasecmp
in GCC which changed with 4.8 and newer builds of GCC.
Added Version to the banner printout. We are currently version 1.00.70.
the 70 is the build number. I'll be incrementing that whenever I do a
build. I'll attempt to tag the source whenever I do a build.