Unused Method Removal Fixes
Fixed tracking of coginit/cognew to get the proper sub index.
Fixed handling of arrays of sub objects or multiple instances of the
same sub object.
Fixed handling of invalid calls due to parsing of embedded strings in
methods as bytecode.
Fixed handling of object names with extensions attached in OBJ section
causing it to think object wasn't used.