UnusedMethodElimination fix + x64 warning fixes
With UnusedMethodElimination, if an object has no methods called, then
the cognew/coginit referenced methods will not be marked as called since
the methods containeding the coginit/cognew are not called. This fixes
something mentioned in issue #21
Added x64 compile options in the VS2017 project files for the
PropellerCompiler.lib, and fixed up a bunch of warnings related to
compiling in x64 mode. Note: openspin itself is still x86 (32bit) and
still compiles the lib in x86(32bit) mode.